app/modal-processor/requirements_railway.txt
google-genai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
fastapi[standard]>=0.115.0
tenacity>=8.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
crawl4ai>=0.4.0
playwright>=1.40.0
uvicorn[standard]>=0.27.0
git+https://github.com/federicodeponte/openpull.git

